Simplify spin polarised case
